+++ Modules/Microsoft365DSC/DSCResources/MSFT_FabricAdminTenantSettings/MSFT_FabricAdminTenantSettings.psm1
@@ -2040,6 +2040,7 @@ function Export-TargetResource
         }
         if ($fixQuotes)
         {
+            $currentDSCBlock = $currentDSCBlock.Replace('`$', '$')
             $currentDSCBlock = $currentDSCBlock.Replace('`', '"')
         }
         $dscContent += $currentDSCBlock